# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "iniscan api document",
        "title": "iniscan API",
        "contact": {},
        "version": "0.0.1"
    },
    "paths": {
        "/explorer/api/address/address_summary": {
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "account summary",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "account"
                ],
                "summary": "account summary",
                "parameters": [
                    {
                        "description": "data",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.AccountSearch"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.AccountSummaryResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "dto.AccountSearch": {
            "type": "object",
            "properties": {
                "address": {
                    "description": "address",
                    "type": "string"
                },
                "contract": {
                    "description": "contract",
                    "type": "string"
                },
                "current": {
                    "type": "integer"
                },
                "pageSize": {
                    "type": "integer"
                }
            }
        },
        "dto.AccountSummaryResponse": {
            "type": "object",
            "properties": {
                "balance": {
                    "type": "string"
                },
                "count": {
                    "type": "integer"
                }
            }
        }
    },
    "securityDefinitions": {
        "Bearer": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}